selenium-autotest-master/requirements.txt
pytest==7.1.2
pytest-html==3.1.1
allure-pytest==2.9.45
PyYAML==6.0
selenium==4.3.0
zmail==0.2.8
requests==2.28.1
